commit do apocalipse - sem testes por falta de ambiente des - mudei a estrutura de rotas e controllers
