+++ app/Models/GestaoImoveisCaixa/ConformidadeContratacao.php
@@ -4,7 +4,7 @@ namespace App\Models\GestaoImoveisCaixa;
 
 use Illuminate\Database\Eloquent\Model;
 
-class ImagemCaixaCecom extends Model
+class ConformidadeContratacao extends Model
 {
     protected $table = 'ADJTBL_imoveisCaixa';
