--- app/Models/GestaoImoveisCaixa/FilaConformidade.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models\GestaoImoveisCaixa;
-
-use Illuminate\Database\Eloquent\Model;
-
-class ConformidadeContratacao extends Model
-{
-    protected $table = 'ADJTBL_imoveisCaixa';
-
-
-}
